# Conflicts:
#	src/Smithbox.Program/Editors/Viewport/Utils/TiledScreenshotTool.cs
#	src/Smithbox.Program/Editors/Viewport/Utils/ViewportUtils.cs

+++ src/Smithbox.Program/Editors/Viewport/Utils/ViewportUtils.cs
@@ -54,11 +54,15 @@ public static class ViewportUtils
         Matrix4x4 fullProj,
         int tileX, int tileY, int multiplier)
     {
-        float N = multiplier;
+        float N = (float)multiplier;
         
         // Calculate center of the tile in clip space [-1, 1]
         float centerX = -1.0f + (2.0f * tileX + 1.0f) / N;
         float centerY = -1.0f + (2.0f * tileY + 1.0f) / N;
+
+        // We want a transformation that maps the tile box to the full [-1, 1] clip space
+        // NewClip = (OldClip - Center) * N
+        // NewClip = OldClip * N - Center * N
         
         Matrix4x4 scale = Matrix4x4.CreateScale(N, N, 1.0f);
         Matrix4x4 translation = Matrix4x4.CreateTranslation(-centerX * N, -centerY * N, 0.0f);
